Extract types
