+++ src/types/bug-area-props.ts
@@ -1,0 +1,5 @@
+import type { Bug } from './bug'
+
+export interface BugAreaProps {
+  bugs: Bug[]
+}
