+++ src/types/bug-card-props.ts
@@ -1,0 +1,7 @@
+import type { Bug } from './bug'
+
+export interface BugCardProps {
+  bug: Bug
+  /** Compact hover preview when true */
+  preview?: boolean
+}
