+++ src/types/bug-crawler-props.ts
@@ -1,0 +1,9 @@
+import type { Bug } from './bug'
+
+export interface BugCrawlerProps {
+  x: number
+  y: number
+  bug: Bug
+  containerWidth: number
+  containerHeight: number
+}
