+++ src/types/bug-trends-chart.ts
@@ -1,0 +1,5 @@
+export interface DataPoint {
+  date: Date
+  created: number
+  resolved: number
+}
