+++ src/types/element-size.ts
@@ -1,0 +1,4 @@
+export interface ElementSize {
+  width: number
+  height: number
+}
