+++ src/types/input-props.ts
@@ -1,0 +1,2 @@
+export interface InputProps
+  extends React.InputHTMLAttributes<HTMLInputElement> {}
