+++ src/types/label-props.ts
@@ -1,0 +1,2 @@
+export interface LabelProps
+  extends React.LabelHTMLAttributes<HTMLLabelElement> {}
