+++ src/types/meta-props.ts
@@ -1,0 +1,6 @@
+export interface MetaProps {
+  title: string
+  description: string
+  image?: string
+  structuredData?: Record<string, unknown>
+}
