+++ src/types/priority-model.ts
@@ -1,0 +1,4 @@
+export interface PriorityModel {
+  weight: number
+  bias: number
+}
