+++ src/types/taskbar-props.ts
@@ -1,0 +1,5 @@
+export interface TaskbarProps {
+  windowTitle: string
+  minimized: boolean
+  onToggle: () => void
+}
