+++ src/types/textarea-props.ts
@@ -1,0 +1,2 @@
+export interface TextareaProps
+  extends React.TextareaHTMLAttributes<HTMLTextAreaElement> {}
